encoding: Decouple adapters from encoding and more
Decouples the queue implementations from the encoding / decoding details
by introducing a new `encoding` package with a JSON implementation and
three interfaces, `Codec`, `Encoder` and `Decoder`.

Adds two new types in `pubsublite`; `Subscription` and `Topic` with a
`String()` and `Validate()` methods to ease their use and concentrate
the formatting of GCP resources.

Signed-off-by: Marc Lopez Rubio <marc5.12@outlook.com>