pubsublite(consumer): default AtMostOnce delivery
Introduce a `Delivery` configuration option for the pubsublite consumer,
allowing the caller to configure the message acknowledgement strategy.

The two supported strategies are: `AtMostOnce` and `AtLeastOnce`. The
former is the current default (if left unspecfied), while the latter is
not exactly at least once (depending on the implementation of the batch
processor).

The configuration options have been documented to avoid confusion.

Signed-off-by: Marc Lopez Rubio <marc5.12@outlook.com>