consumer(kafka): Concurrently consume partitions
Updates the Kafka consumer to consume from partitions via dedicated go-
routines. Each assigned partition for the consumer in the consumer group
will be consumed concurrently.

A `consumer` and `partitionConsumer` types are introduced, the former
manages the assigned and revoked / lost partitions, and ensure that the
latter (`partitionConsumer`) is created on assignment of new partitions,
and stopped on revoked / lost partitions.

Signed-off-by: Marc Lopez Rubio <marc5.12@outlook.com>