kafka(sasl): Abstract sasl.Mechanism via packages
Abstracts the underlying `sasl.Mechanism` interface (originating from
`franz-go`) in favor of dedicated `sasl/<type>` packages. A new package
`sasl/plain` is introduced, and `sasl/aws` is introduced as a module to
avoid importing the AWS SDK directly in AWS queue for cases when the
importer may not want or need to use the AWS MSK IAM sasl type.

Signed-off-by: Marc Lopez Rubio <marc5.12@outlook.com>