pubsublite(producer): Allow async producing (#59)
Adds a new `pubsublite.ProducerConfig.Sync` config option which allows
callers to configure the nature of record publishing. By default, it
will use async producing, which offers infinitely more performance.

There is some work to do, still if the producing a message fails with an
error. This is a non-recoverable error and when it occurs, it renders
the current topic `pscompat.PublisherClient` useless, which means that
we should attempt to re-create the `pscompat.PublisherClient` records
can be produced. It is unclear which kinds of errors would trigger this
case, but that's what the comments in the pubsublite library explain.

Closes #54

Signed-off-by: Marc Lopez Rubio <marc5.12@outlook.com>