systemtest: Add initial framework and tests
Adds the initial framework for `systemtest` as a new go module under
`systemtest/`. The current strategy is to rely on Terraform for any
infrastructure provisioning since that allows us to re-use many of the
existing modules and have a pluggable interface.

The tests will initially provision a Kafka cluster in a k8s cluster of
choice, with a set of topics, and the same for GCP PubSub Lite.

To be able to access the Kafka cluster, we do two things:

- Use the `kubectl port-forward` command to forward the traffic to the
k8s kafka service.
- Patch the `kgo.Dialer` with a DNS resolution that maps the kafka addr
to the localhost mapping.

Closes #92.
Part of #93.

Signed-off-by: Marc Lopez Rubio <marc5.12@outlook.com>