systemtest: Add initial framework and tests (#99)
Adds the initial framework for `systemtest` as a new go module under
`systemtest/`. The current strategy is to rely on Terraform for any
infrastructure provisioning since that allows us to re-use many of the
existing modules and have a pluggable interface.

The tests will initially provision a Kafka cluster in a k8s cluster of
choice, with a set of topics, and the same for GCP PubSub Lite.

To be able to access the Kafka cluster, we do two things:

- Use the `kubectl port-forward` command to forward the traffic to the
k8s kafka service.
- Patch the `kgo.Dialer` with a DNS resolution that maps the kafka addr
to the localhost mapping.

The containing kafka terraform must accept at least three variables:

- `namespace (string)`: Namespace where the kafka cluster will be
created
- `name (string)`: Name for the Kafka cluster.
- `Topics (list(string))`: List of topics to create.

And it is expected to output at least two outputs:

- `deployment_type (string)`: Type of Kafka deployment. The value of
this output will determine some of the behavior for the ProvisionKafka
function. If `deployment_type` output is `"k8s"`, `kubectl port-forward`
will be run targeting the `PortForwardResource` with
`PortForwardMapping`. Any other terraform types could be "other" or
export the provider where they've been provisioned, like `EventHubs`,
`MSK`, etc.
- `kafka_brokers (list(string))`: The list of kafka brokers to use for
clients to connect to the Kafka cluster. The hosts must be reachable and
resolvable from the machine that runs the system tests.

Closes #92.
Part of #93.

---------

Signed-off-by: Marc Lopez Rubio <marc5.12@outlook.com>