Sleep in producer test to reduce the risk of flaky test (#108)
franz-go runs the hooks to finish events in an unchecked goroutine:


https://github.com/twmb/franz-go/blob/master/pkg/kgo/producer.go#L436-L440

This method calls `finishPromises`, which calls
`OnProduceRecordUnbuffered`, which is the hook that kotel uses to end
the span.

Because of that, the check on the number of created spans is flaky:
https://github.com/elastic/apm-queue/actions/runs/4763014463/jobs/8466352708

We don't have any way to know whether kgo has finished running the
goroutine. So by sleeping for a millisecond, we reduce the risk of the
flaky issue happening.
If kgo took more than 1ms to process the hooks, that seems like
something which would be a performance issue to investigate anyway.