kafka(consumer): Improve `AtLeastOnce` handling
Improves how `AtLeastOnceDelivery` is handled by the kafka consumer.
Previously, the consumer could discard up to `cfg.MaxPollRecords`, since
it incorrectly assumed that unacknowledged records would be re-delivered
by Kafka if not committed within a certain amount of time.

Now, it discards any records which can't be processed, but continues
processing the rest of the records in the fetch. It can still result in
some data loss, and is at the mercy of the failures returned by the
processor, but it's much better than previously.

Additionally, adds a few tests which cover both `AtLeastOnceDelivery`
and `AtMostOnceDelivery`, and a `TestGracefulShutdown` test.

Partially addresses #118
Part of #123

Signed-off-by: Marc Lopez Rubio <marc5.12@outlook.com>