kafka(consumer): Fix At Least Once delivery commit
Update the partition consumer to skip committing offsets of records that
have failed processing when AtLeastOnceDelivery is used. May not cause
re-processing of records if newer records are produced to the partition,
unless the consumer crashes or the partition is reassigned to a another
consumer in the consumer group. This is still preferrable to committing
the offset of a failed record, given that AtLeastOnceDelivery is set.

Signed-off-by: Marc Lopez Rubio <marc5.12@outlook.com>