Add Manager types, and TopicCreator interface
Add types for managing the creation of Kafka topics,
and the creation of Pub/Sub Lite reservations, topics,
and subscriptions.

pubsublite.ConsumerConfig has been updated to replace
Topics with Subscriptions. A topic may have multiple
subscriptions, roughly equivalent to consumer groups
in Kafka.

The system tests will be updated to use these in a
followup PR.