consumer(kafka): Use record.Context in processor
Use each of the record's context in `processor.ProcessBatch` in order to
propagate the record's headers in the context. Introduce a `combinedCtx`
which uses the deadline of the first context, but redirects the calls to
`ctx.Value` to the second context.

This allows the `kgo.Client.Close()` call to cancel the first context
and trigger shutdown to stop record processing.

Signed-off-by: Marc Lopez Rubio <marc5.12@outlook.com>