pubsublite(publisher): Record metrics, add testing
This PR is the counterpart of #181 for the pubsublite package, with a
bit of refactoring that was needed in order to be able test the metrics.

Introduces a new package `pubsubabs` which provides the necessary means
to abstract the types in the `pubsub` package. Two new interfaces are
introduced; `PublisherClient` implements the `pubsub.PublisherClient`,
and `Publisher`, which returns a `PublishResult` interface that can be
used to assert that our business logic works as expected in unit tests.

This allows us to start writing unit tests for the pubsublite.Publisher.

`telemetry.Producer` now implements the `pubsubabs.Publisher` interface
and has been refactored from a function to a type.

`producer.messages.produced` and `produced.messages.errored` metrics are
now recorded and reported using the configured OTel meter provider.

Signed-off-by: Marc Lopez Rubio <marc5.12@outlook.com>