refactor tests to remove flakiness
I refactored the test to:
- identify the correct ScopeMetrics to target for assertions
- use separate readers to avoid compounding collected metrics
- extracted common logic to helper functions
- test only the specific attributes we are intested in without using
go-cmp

I decided to remove go-cmp comparator as it was not providing much value
once I refactored the test to target the specific fields I was
interested in.

As these test are not flaky anymore I also removed the metrics build
tag, as is no more useful.