Introduce `Record` and `Processor`, remove `Codec`
This change transforms this module into a generic consume/produce to
Kafka or PubSub Lite.
It decouples record encoding from producing, decoding from consuming and
removes the inter-module dependency on `apm-data`.

Introduces a new top level `apmqueue.Record`, wrapping a slice of bytes
(`[]byte`) with an `apmqueue.Topic`. The producer also interface is also
changed to:

    Produce(context.Context, ...Record) error

A new top level `apmqueue.Producer` is also introduced which is accepted
by the two consumer implementations.

    Process(context.Context, ...Record) error

Last it removes the `cmd.loadgen` since it's not as useful since any
data can be produced to the queue, not only `model.Batch`.

`codec/json` is also removed since it's been moved to `apm-data`.

Signed-off-by: Marc Lopez Rubio <marc5.12@outlook.com>