kafka: introduce config file option
Introduce CommonConfig.ConfigFile, which
should point to a YAML file with configuration.
The filename may also be specified with the
environment variable $KAFKA_CONFIG_FILE.

The config file can include:
 - bootstrap.servers
 - sasl.mechanism
 - sasl.username
 - sasl.password

If a config file is specified, a hook will be
added to kgo.Client to reload seed brokers from
the file when a broker connection fails. Thus,
changes to the file will be picked up when the
old broker addresses begin failing.

If `sasl.mechanism` is set to PLAIN, or
`sasl.username` is specified, then SASL/PLAIN
is configured and will reload the username and
password from the file on each authentication
attempt. Thus, changes to the credentials will
be picked up when connections are recreated.

Presence of the AWS_ACCESS_KEY_ID environment
variable no longer configures SASL/AWS_MSK_IAM
automatically. It is now necessary to specify
`KAFKA_SASL_MECHANISM=AWS_MSK_IAM`, or to specify
`sasl.mechanism: AWS_MSK_IAM` in the config file.
When this is set, we now use the AWS SDK to
confgure AWS_MSK_IAM; changes to credentials
files will be picked up on reauthentication.