kafka: add topic namespacing (#246)
If `CommonConfig.Namespace` is non-empty, a prefix (filter) will be
added to all topics. This enables the creation of a topic with the same
logical name within two different namespaces. The namespace will be
stripped from returned topic names, and will be added as a separate
attribute to metrics and logs.

Closes https://github.com/elastic/apm-queue/issues/224