use context cancellation to handle os signals
Use signal.NotifyContext to react to os.Interrupt and syscall.SIGTERM.
Handle context cancellation in wait loop and production loop.

By cancelling the context we can reuse the same teardown logic, without
triggering it in two different places.

The main context was also used in teardown topic deletion, which
resulted in it being skipped on context cancellation. It now uses a
different context.

Co-authored-by: Andrew Wilkins <axwalk@gmail.com>