don't timeout on each wait loop iteration
Remove timeout on wait loop iteration.

The timeout was added to stop any long running operation within the wait
loop taking more than the ticker time.
This is considered not necessary - as there are no long running
operations in the loop, and had a bug in context cancellation, which was
deferred to the end of the main function instead of the wait loop.

Co-authored-by: Andrew Wilkins <axwalk@gmail.com>