test: refactor unit tests to reduce flakyness (#297)
Tested locally by running tests with `-count=<big>` and was able to
reduce flakyness significantly.

## Changes:

### TestConsumerDelivery:

Instead of waiting for `consumer.Close` to complete before proceeding,
call `consumer.Close` from the same goroutine and wait for
`consumer.Run` to exit. This ensure that the first consumer has exited.

Call the `cancel` function inside the processor to avoid a small race
between the time the processor finishes and the time `cancel` was
called.

### TestNewProducerBasic

Set `MaxBufferedRecords` to 0 to avoid context cancelled because the
consumer is waiting on the buffered records. (another option would be to
close the producer to force a flush).