Wrap errors with additional information (#321)
`apm-queue` is a library, which leads to these errors surfacing in other
code.

Debugging errors from libraries is painful without clear indication of
where the error comes from.

To help with debugging this PR wrap errors with additional details.

One change that affects an `err` variable but is not an addition is the
removal of a `return err` in `pubsublite/topiccreator.go`. The function
handles the case `err != nil` separately so the only case in which that
`return err` was executed was on `err = nil`. To increase clarity I
changed it to directly `return nil`.