test: update AWS_SHARED_CREDENTIALS_FILE to point to a file
test was failing because AWS_SHARED_CREDENTIALS_FILE was pointing to a directory
which caused the config lodaing code to fail with an error.
The value of the variable is not used so the file doesn't need to exist
but pointing to an existing directory leads to failure.
Update the code to use a file.