kafka: Remove partition consumer goroutines
Removes the need for dedicated goroutines to be spun when the consumer
receives its partition assignments. Instead, it uses an `errgroup.Group`
with a `MaxProcessorConcurrency` (default: `1`), which behaves in the
same manner.

This approach's major advantage is that it reduces the number of running
goroutines, and it's noticeable as the partition count increases.

Without this patch, the consumer throughput plateaued after certain
throughput is reached.

Signed-off-by: Marc Lopez Rubio <marc5.12@outlook.com>