test: fix flaky TestConsumerDelivery
Processing happens in a separate goroutine so even if we cancel the context as
soon as we start processing the first batch of records a second fetch might
have completed already.
The other scenario is that the context is canceled before the second fetch starts
creating a +- tc.maxPollRecords variance in errored records and potentially
causing extra records in processed record counter.
Update the test to ignore extra records processed after the context has been
canceled and account for an extra tc.maxPollRecords records when asserting
processed records.
Make the test more deterministic and assert for equality instead of comparing >=.

This issue only happens if tc.initialRecords != tc.maxPollRecords.

Mostly affecting 30_produced_2_poll_AMOD.