Kafka: Support specifying client certificates (#636)
Adds support for Kafka to use client certificates based on two new env
variables `KAFKA_TLS_CERT_PATH` and `KAFKA_TLS_KEY_PATH`. The current
implementation supports hot reloading by checking if the file write time
is greater than the last time it was read.

Most of the changes are in the form of unit tests.

---------

Signed-off-by: Marc Lopez Rubio <marc5.12@outlook.com>