feat(kafka): Introduce `LogCompactedConsumer`
Introduces a new consumer type, `LogCompactedConsumer`, specifically
designed for consuming from Kafka topics with log compaction enabled.

Its intended use case is for applications that need to maintain a whole,
up-to-date snapshot of the state represented in a log-compacted topic.
Unlike a standard consumer that might start processing from the latest
offset, the `LogCompactedConsumer` always starts from the beginning of
each partition. Ensuring all records are processed in order to build a
complete view of the data.

The LogCompactedConsumer `Health()` method is tied to an initial "sync".
When a new consumer instance is started, especially in a topic with a
significant backlog, it will begin consuming all messages from offset 0
(or latest available since deletion/compaction) for all its assigned
partitions. During this period, the `Healthy()` method blocks until the
consumer has fully synced or the context is done (an error is returned)
indicating that the consumer has not yet reached the high-water mark for
all partitions of the log compacted topic.

This _may_ be crucial for applications that need to be fully initialized
with the topic's data before they can serve traffic or be considered
"ready". For example, in a Kubernetes environment, the readiness probe
can use the `Healthy()` method to prevent a pod from receiving traffic
until it has fully synchronized its state from the log-compacted topic.

Errors returned by the configured processor are treated as non-fatal;
they are logged, but the consumer continues to process subsequent
records. This design choice is based on the nature of log-compacted
topics, where the primary goal is to consume the latest state for each
key, and a transient failure in processing one record should not prevent
the consumer from receiving future updates.

Signed-off-by: Marc Lopez Rubio <marc5.12@outlook.com>